Challenge and verify
Completed the documentation for the pointers and structures challenge, including the challenge description. The verify script was mostly copied from the arrays-and-slices one, since the core of the challenge is the same. The new template contains the necessary modifications as well. There was a 3 character change to the array description to change 'topic' to 'topic(s)' to keep elements standard across challenges.