Feature/pointers (#30)
* Pointers

Created a new challenge for pointers and structs, and modified the intro description to allow for more variety of challenges instead of just arrays, slices, runes, and constants.

* Challenge and verify

Completed the documentation for the pointers and structures challenge, including the challenge description. The verify script was mostly copied from the arrays-and-slices one, since the core of the challenge is the same. The new template contains the necessary modifications as well. There was a 3 character change to the array description to change 'topic' to 'topic(s)' to keep elements standard across challenges.

* Update Description

Discovered the formulas copied from arrays and slices do not render in the browser, so updated with unicode characters.

* Updated Description

Minor fix.

* Update/slice description (#29)

* Update Description

Discovered the formulas do not render in the browser, so updated with unicode characters.

* Update Description

Minor fix.

---------

Co-authored-by: ARL <awollman@madren.org>

* Fixed missing '\n' in first print statement to enable verifer script.

---------

Co-authored-by: ARL <awollman@madren.org>