Feature/classes (#31)
* Classes

While not strictly accurate, this work begins the 'classes' challenge. Go doesn't have classes; they use receiver methods to implement functionality specific to structs and interfaces to standardize access/use.

* Challenge development

Started developing challenge text and prototyping a solution.

* Challenge development

Completed the description document, created the challenge, and copied the main.go template file to the folder. Need to complete the verify script and develop a solution file to check against.

* Verifier script

Completed the go template file, verifier file, and description document.

* Modifications

Fixed part of the verifier script and a type-o in the description.

---------

Co-authored-by: ARL <awollman@madren.org>